This is a massive round of changes to make the cross-compile 64-bit clean.  Really, there's one change that had to be done:  All references to long had to be transformed to int32_t, and all changes to unsigned long had to be transformed to uint32_t.
Note, ultimately, the reason these changes are event needed is because
some tools in utils (makedefs, lev_comp, dlb, etc) use NetHack structs
to write out binary data which must then be readable by the game itself.
But if you build those utilities on a 64-bit platform, and longs are
used to write out that data, then you'll end up with 64-bit values on
write, and 32-bit values on the NDS at read time.

Now, this does mean I probably could've tried to minimize these changes
and isolate them to just those places where said structs are written
out and read back.  Unfortunately, the ripple affect of those changes
was big enough that I didn't think it was worth the trouble to try and
surgically fix things.  Plus, a surgical approach would've inevitably
been more error-prone.

So instead, I did the mass change to all include and src files.
Hopefully I got them all (at least the game starts up and runs, now).

+++ include/decl.h
@@ -167,7 +167,7 @@ E const char ynchars[];
 E const char ynqchars[];
 E const char ynaqchars[];
 E const char ynNaqchars[];
-E NEARDATA long yn_number;
+E NEARDATA int32_t yn_number;
 
 E const char disclosure_options[];
 
@@ -181,7 +181,7 @@ E NEARDATA int killer_format;
 E const char *killer;
 E const char *delayed_killer;
 #ifdef GOLDOBJ
-E long done_money;
+E int32_t done_money;
 #endif
 E char killer_buf[BUFSZ];
 E const char *configfile;
@@ -206,12 +206,12 @@ E NEARDATA struct dig_info {		/* apply.c, hack.c */
 	int	effort;
 	d_level level;
 	coord	pos;
-	long lastdigtime;
+	int32_t lastdigtime;
 	boolean down, chew, warned, quiet;
 } digging;
 
-E NEARDATA long moves, monstermoves;
-E NEARDATA long wailmsg;
+E NEARDATA int32_t moves, monstermoves;
+E NEARDATA int32_t wailmsg;
 
 E NEARDATA boolean in_mklev;
 E NEARDATA boolean stoned;
